Ion

Ion.COMMAND_ENTRY_DOC_STUB

Ion.clone

Ion.comonicon_install

Ion.comonicon_install_path

Ion.compat

Ion.release

Ion.COMMAND_ENTRY_DOC_STUB — Constant

CLI&REPL toolkit for managing/developing Julia project & packages.

Ion.clone — Method

clone(package_or_url::String, to::String=pwd(); force::Bool=false)

Clone a package or remote URL to a local directory.

Intro

This command helps you clone a remote package repo from github. The corresponding package or repo name will automatically remove the .jl extension if it has one, e.g https://github.com/JuliaLang/Example.jl will be downloaded as Example. And when cloning a repo that has no write access, it will prompt to ask for a fork.

Arguments

package_or_url: package name or URL.

to: local directory to download the git repo to.

Flags

-f,--force: force overwrite existing path.

-y,--yes: skip prompt with answer yes.

Ion.comonicon_install — Method

comonicon_install(;kwargs...)

Install the CLI manually. This will use the default configuration in Comonicon.toml, if it exists. For more detailed reference, please refer to Comonicon documentation.

Ion.comonicon_install_path — Method

comonicon_install_path(;[yes=false])

Install the PATH and FPATH to your shell configuration file. You can use comonicon_install_path(;yes=true) to skip interactive prompt. For more detailed reference, please refer to Comonicon documentation.

Ion.compat — Method

compat(version_spec::String="auto", path_to_project::String=pwd(); package::String="", overwrite::Bool=false)

create and update compat in Project.toml via command line.

Intro

By default the auto update the compat to latest non-breaking version by semantic version definition (latest minor version for 0.x.y, latest major version for other case). One can select only updating specific packages by using the package option.

Arguments

version_spec: version_spec to be compatible with, can be patch, minor, major or auto.

path_to_project: path to the project, default is the current working directory.

Options

-p,--package=<name>: package selector, use this option to update specified package compat only.

Flags

--overwrite: force update all compat by overwrite old compat.

Ion.release — Function

release(version_spec[, path=pwd()]; kw...)

Release a new version for a package, e.g

julia> pwd()
"/home/roger/julia/Comonicon"

julia> Ion.release(patch) # will release a new patch version on github

Arguments

version_spec: required, semantic version bump (patch, minor, major, current) or a specific version number to release (VersionNumber or valid version string).

path: optional, path to the pacakge to release, default is pwd().

Keyword Arguments

registry: optional, name of the registry to register, if not specified, will lookup if the package has been registered in local registries.

branch: optional, branch to register, default is current branch.

note: optional, release note.